+++ components/breadcrumb/test/breadcrumb.test.js
@@ -1,0 +1,21 @@
+import '../vwc-breadcrumb.js';
+import {
+	waitNextTask,
+	textToDomToParent,
+	isolatedElementsCreation
+} from '../../../test/test-helpers.js';
+import { chaiDomDiff } from '@open-wc/semantic-dom-diff';
+
+chai.use(chaiDomDiff);
+
+const COMPONENT_NAME = 'vwc-breadcrumb';
+
+describe('Breadcrumb', () => {
+	let addElement = isolatedElementsCreation();
+
+	it(`${COMPONENT_NAME} is defined as a custom element`, async () => {
+		assert.exists(
+			customElements.get(COMPONENT_NAME)
+		);
+	});
+});