+++ components/card/src/vwc-card.ts
@@ -1,17 +1,22 @@
 import {
 	customElement, html, LitElement,
 } from 'lit-element';
+import { nothing } from 'lit-html';
 import { style } from './vwc-card.css.js';
 import { property } from 'lit-element/lib/decorators.js';
 import { classMap } from 'lit-html/directives/class-map.js';
 import '@vonage/vwc-button';
 import '@vonage/vwc-icon';
+import '@vonage/vwc-elevation';
 
 declare global {
 	interface HTMLElementTagNameMap {
 		'vwc-card': VWCCard;
 	}
 }
+
+const elevationSets = ['0' , '2', '4', '8', '12' , '16' , '24'];
+export type IndicatorElevationSets = typeof elevationSets;
 
 /**
  * @cssprop [--title-line-clamp] defines the number of lines presented before trim + ellipsis in the card title
@@ -39,72 +44,91 @@ export class VWCCard extends LitElement {
 
 	@property({
 		reflect: true,
-		attribute: 'header-icon',
+		attribute: 'icon',
 		type: String
 	})
-		headerIcon: string | null = null;
+		icon: string | null = null;
 
 	@property({
 		reflect: true,
-		attribute: 'supporting-text',
+		attribute: 'text',
 		type: String
 	})
-		supportingText: string | undefined;
+		text: string | undefined;
 
-	private headerIconSlottedItems?: Node[];
-	private shouldShowActionsSlot: boolean | undefined;
+	@property({
+		reflect: false,
+		attribute: 'elevation',
+		type: String
+	})
+		elevation: IndicatorElevationSets[number] = '4';
+
+	private IconSlottedItems?: Node[];
+	#shouldShowFooterSlot: boolean | undefined;
 
 	private get headerContentExists(): boolean {
-		return Boolean(this.heading || this.subtitle || this.headerIcon || this.headerIconSlottedItems?.length);
+		return Boolean(this.heading || this.subtitle || this.icon || this.IconSlottedItems?.length);
 	}
 
 	private get headerClass(): string {
 		return (this.headerContentExists) ? '' : 'no-content';
 	}
 
+
 	protected override render(): unknown {
-		const actionsClassMap = {
-			'no-content': !(this.shouldShowActionsSlot)
+		const footerClassMap = {
+			'no-content': !(this.#shouldShowFooterSlot)
 		};
 		return html`
-			<div class="vwc-card">
-				<div class="vwc-card-media">
-					<slot name="media"></slot>
-				</div>
-				<div class="vwc-card-info">
-					${this.renderHeader()}
-					<div class="vwc-card-supportText">
-							${this.supportingText ? this.supportingText : ''}
+			<vwc-elevation .dp=${this.elevation}>
+			<!-- there are 2 wrapper due to a safari bug failing 'filter: drop-shadow'
+			from rendering shadow on an element with 'overflow: hidden' -->
+				<div class="vwc-card">
+					<div class="vwc-card-container">
+						<div class="vwc-card-media">
+							<slot name="media"></slot>
+						</div>
+						<slot name="main">
+							<div class="vwc-card-content">
+								<div class="vwc-card-wrapper">
+									${this.renderHeader()}
+									<slot name="meta"></slot>
+								</div>
+								<div class="vwc-card-text">
+									${this.text ? this.text : nothing}
+								</div>
+							</div>
+						</slot>
+						<div class="vwc-card-footer ${classMap(footerClassMap)}">
+							<slot name="footer" @slotchange="${this.footerSlotChanged}"></slot>
+						</div>
 					</div>
-					<div class="vwc-card-actions ${classMap(actionsClassMap)}">
-							<slot name="actions" @slotchange="${this.actionsSlotChanged}"></slot>
-					</div>
 				</div>
-			</div>
+			</vwc-elevation>
 		`;
 	}
 
 	private renderHeader() {
 		return html`
-			<header class="${this.headerClass}">
-				<div class="vwc-card-header">
-					<slot name="graphics" @slotchange="${this.graphicsSlotChanged}">
-						${this.headerIcon ? this.renderIcon() : ''}
-					</slot>
+			<header class="vwc-card-header ${this.headerClass}">
+				<slot name="graphic" @slotchange="${this.graphicSlotChanged}">
+					${this.icon ? this.renderIcon() : ''}
+				</slot>
+				<div>
 					<div class="vwc-card-title">${this.heading}</div>
+					<div class="vwc-card-subtitle">${this.subtitle}</div>
 				</div>
-				<div class="vwc-card-subtitle">${this.subtitle}</div>
 			</header>`;
 	}
 
 	private renderIcon() {
-		return html`<vwc-icon class="header-icon" inline type="${this.headerIcon}"></vwc-icon>`;
+		return html`<vwc-icon class="icon" inline type="${this.icon}"></vwc-icon>`;
 	}
 
-	private graphicsSlotChanged() {
+	private graphicSlotChanged() {
 		const headerElement = this.shadowRoot?.querySelector('header');
-		const slot = headerElement?.querySelector('slot[name="graphics"]') as HTMLSlotElement;
-		this.headerIconSlottedItems = slot.assignedNodes();
+		const slot = headerElement?.querySelector('slot[name="graphic"]') as HTMLSlotElement;
+		this.IconSlottedItems = slot.assignedNodes();
 		if (this.headerContentExists) {
 			headerElement?.classList.remove('no-content');
 		} else {
@@ -112,9 +136,9 @@ export class VWCCard extends LitElement {
 		}
 	}
 
-	private actionsSlotChanged(): void {
-		const slot = this.shadowRoot?.querySelector('slot[name="actions"]') as HTMLSlotElement;
-		this.shouldShowActionsSlot = Boolean(slot.assignedNodes().length);
+	private footerSlotChanged(): void {
+		const slot = this.shadowRoot?.querySelector('slot[name="footer"]') as HTMLSlotElement;
+		this.#shouldShowFooterSlot = Boolean(slot.assignedNodes().length);
 		this.requestUpdate();
 	}
 }