+++ components/tags/src/vwc-tag-base.ts
@@ -8,6 +8,7 @@ import type { ClassInfo } from 'lit-html/directives/class-map.js';
 import {
 	LitElement, html, property, TemplateResult, queryAsync, state, query, eventOptions
 } from 'lit-element';
+import { nothing } from 'lit-html';
 
 
 type TagConnotation = Extract<
@@ -43,6 +44,12 @@ export class VWCTagBase extends LitElement {
 
 	@property({ type: Boolean, reflect: true })
 		selectable = false;
+
+	@property({ type: Boolean, reflect: true })
+		removable = false;
+
+	@property({ type: Boolean, reflect: true })
+		removeEventOnly = false;
 
 	@state() protected shouldRenderRipple = false;
 
@@ -120,11 +127,25 @@ export class VWCTagBase extends LitElement {
 		</span>`;
 	}
 
+	protected removeTag(): void {
+		this.dispatchEvent(new CustomEvent('remove-tag', {detail: this}));
+		if (!this.removeEventOnly) {
+			this.remove();
+		}
+	}
+
+	protected renderRemoveButton(): TemplateResult {
+		return html`<button class="remove-button" @click="${()=> this.removeTag()}">
+			${this.renderIcon('close-line')}
+		</button>`;
+	}
+
 	override render(): TemplateResult {
 		return this.selectable
 			? this.renderTagSelectable()
 			: html`<span class="vwc-tag ${classMap(this.getRenderClasses())}">
 			${this.text}
+			${this.removable ? this.renderRemoveButton() : nothing}
 		</span>`;
 	}